Add new methods in ProductRepositoryImpl for find product by id, and get all products as list

+++ spring-mvc/src/main/java/com/springmvclesson3/springmvc/domain/Product.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class Product {
     private long id;
     private String title;
-    private double cost;
+    private int cost;
 }
